Move XMLConfig to its own file
