+++ python/astra/src/PythonPluginAlgorithm.cpp
@@ -29,6 +29,7 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 
 #include "PythonPluginAlgorithm.h"
 
+#include "astra/XMLConfig.h"
 #include "astra/Logging.h"
 #include "astra/Utilities.h"
 #include <iostream>
