+++ src/ConeVecProjectionGeometry3D.cpp
@@ -26,6 +26,8 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "astra/ConeVecProjectionGeometry3D.h"
+
+#include "astra/XMLConfig.h"
 #include "astra/Utilities.h"
 #include "astra/Logging.h"
