+++ src/FanFlatProjectionGeometry2D.cpp
@@ -28,7 +28,7 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 #include "astra/FanFlatProjectionGeometry2D.h"
 
 #include "astra/GeometryUtil2D.h"
-
+#include "astra/XMLConfig.h"
 #include "astra/Logging.h"
 
 #include <cstring>
