+++ src/ParallelVecProjectionGeometry2D.cpp
@@ -27,6 +27,7 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 
 #include "astra/ParallelVecProjectionGeometry2D.h"
 
+#include "astra/XMLConfig.h"
 #include "astra/Logging.h"
 
 #include <cstring>
