+++ src/ParallelVecProjectionGeometry3D.cpp
@@ -27,6 +27,7 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 
 #include "astra/ParallelVecProjectionGeometry3D.h"
 #include "astra/Utilities.h"
+#include "astra/XMLConfig.h"
 #include "astra/Logging.h"
 
 #include <cstring>
