+++ src/VolumeGeometry3D.cpp
@@ -26,6 +26,7 @@ along with the ASTRA Toolbox. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "astra/VolumeGeometry3D.h"
+#include "astra/XMLConfig.h"
 #include "astra/Logging.h"
 
 namespace astra
